refactor: Phase 3 modular architecture — routes/, services/, state.py
- Extract all /api/* route handlers to routes/api.py (Blueprint)
- Extract index route to routes/views.py (Blueprint)
- Create state.py: shared _clients list + adapter name cache
- Create services/bluetooth.py: bt_remove_device, persist_device_enabled, is_audio_device
- Slim web_interface.py to ~60 lines: app init, WSGI middleware, blueprint registration, main()
- Update Dockerfile: COPY routes/, services/, state.py
- Fix state.py: str|None type hint → forward-ref string for Python 3.9 compat
