+++ services/__init__.py
@@ -1,0 +1,4 @@
+"""Services package for sendspin-bt-bridge."""
+from services.bluetooth import bt_remove_device, persist_device_enabled, is_audio_device
+
+__all__ = ['bt_remove_device', 'persist_device_enabled', 'is_audio_device']
